Use uint32_t instead of size_t

+++ automaton/base/cell.hpp
@@ -1,14 +1,14 @@
 #pragma once
 
-#include <stddef.h>
+#include <cstdint>
 
 namespace automaton {
 
 struct base_cell {
-    size_t row;
-    size_t col;
+    uint32_t row;
+    uint32_t col;
 
-    base_cell(size_t row, size_t col) : row(row), col(col) {}
+    base_cell(uint32_t row, uint32_t col) : row(row), col(col) {}
     virtual bool operator<(const base_cell& other) const {
         if (row == other.row) return col < other.col;
         // return backwards to put upper row in std::set first
